Chore: Error handling (#11)
* Chore: Add error handling.
* Refactor: Improve the error handling.
* Refactor: Improve the code structure.

+++ src/product_repository.py
@@ -3,6 +3,7 @@ import os
 from argparse import Namespace
 from typing import Optional
 from conflog import logger
+from datetime import date
 
 from database_context_manager import DatabaseContextManager
 import products
@@ -46,25 +47,29 @@ class ProductRepository(Repository):
         self.host = host or os.environ.get("DB_HOST")
         self.port = port or int(os.environ.get("DB_PORT"))
 
-    def get(self, product_id: int) -> products.Product:
+    def get(self, product_id: int) -> Optional[products.Product]:
         with DatabaseContextManager(database=self.database, user=self.user, password=self.password,
                                     host=self.host, port=self.port) as cursor:
 
-            logger.info(f"Getting product...",
+            logger.info("Getting product...",
                         extra={"extra_parameters": {"product_id": product_id}})
 
             cursor.execute(f"SELECT category, name, expiry_date, quantity FROM products "
                            f"WHERE product_id = '{product_id}';")
 
-            logger.info(f"Getting product is completed.")
-
             products_list = cursor.fetchall()
 
-            product_tuple = products_list[0]
-            product_class = get_product_class(product_tuple[0])
-            product = product_class(product_tuple[1],
-                                    product_tuple[2],
-                                    product_tuple[3])
+            if products_list:
+                product_tuple = products_list[0]
+                product_class = get_product_class(product_tuple[0])
+                product = product_class(product_tuple[1],
+                                        product_tuple[2],
+                                        product_tuple[3])
+                logger.info("Getting product is completed.")
+            else:
+                product = None
+                logger.error("Product not found.",
+                             extra={"extra_parameters": {"product_id": product_id}})
 
             return product
 
@@ -72,7 +77,7 @@ class ProductRepository(Repository):
         with DatabaseContextManager(database=self.database, user=self.user, password=self.password,
                                     host=self.host, port=self.port) as cursor:
 
-            logger.info(f"Searching products...",
+            logger.info("Searching products...",
                         extra={"extra_parameters": {"conditions": conditions, "columns": columns, "limit": limit}})
 
             query = f"SELECT {columns} FROM products"
@@ -85,7 +90,6 @@ class ProductRepository(Repository):
                                                 f" '{attribute_value.get("value")}'")
                     else:
                         where_conditions.append(f"{attribute_name} = '{attribute_value}'")
-
                 query += " WHERE " + " AND ".join(where_conditions)
 
             if limit:
@@ -93,15 +97,25 @@ class ProductRepository(Repository):
 
             cursor.execute(query)
 
-            logger.info(f"Search completed.")
+            logger.info("Search completed.")
 
             return cursor.fetchall()
+
+    def get_by_name_and_date(self, name: str, expiry_date: date) -> list[tuple]:
+        product = self.search(conditions={"name": name, "expiry_date": expiry_date})
+
+        if len(product) > 1:
+            logger.warning("There are more than 1 rows in the database storing the same product.",
+                           extra={"extra_parameters": {
+                               "name": name, "expiry_date": expiry_date}})
+
+        return product
 
     def add(self, product: products.Product) -> None:
         with DatabaseContextManager(database=self.database, user=self.user, password=self.password,
                                     host=self.host, port=self.port) as cursor:
 
-            logger.info(f"Adding product...",
+            logger.info("Adding product...",
                         extra={"extra_parameters": {"category": product.CATEGORY, "name": product.name,
                                                     "expiry_date": product.expiry_date, "quantity": product.quantity}})
 
@@ -109,31 +123,31 @@ class ProductRepository(Repository):
                            f"VALUES('{product.CATEGORY}', '{product.name}',"
                            f"'{product.expiry_date}', {product.quantity});")
 
-            logger.info(f"Adding completed.")
+            logger.info("Adding completed.")
 
     def update(self, product: products.Product) -> None:
         with DatabaseContextManager(database=self.database, user=self.user, password=self.password,
                                     host=self.host, port=self.port) as cursor:
 
-            logger.info(f"Updating quantity of the product...",
+            logger.info("Updating quantity of the product...",
                         extra={"extra_parameters": {"category": product.CATEGORY, "name": product.name,
                                                     "expiry_date": product.expiry_date, "quantity": product.quantity}})
 
             cursor.execute(f"UPDATE products SET quantity = {product.quantity} "
                            f"WHERE name = '{product.name}' AND expiry_date = '{product.expiry_date}';")
 
-            logger.info(f"Updating completed.")
+            logger.info("Updating completed.")
 
     def remove(self, product_id: int) -> None:
         with DatabaseContextManager(database=self.database, user=self.user, password=self.password,
                                     host=self.host, port=self.port) as cursor:
 
-            logger.info(f"Removing product...",
+            logger.info("Removing product...",
                         extra={"extra_parameters": {"product_id": product_id}})
 
             cursor.execute(f"DELETE FROM products WHERE product_id = '{product_id}';")
 
-            logger.info(f"Removing completed.")
+            logger.info("Removing completed.")
 
     def count(self) -> int:
         with DatabaseContextManager(database=self.database, user=self.user, password=self.password,
@@ -143,7 +157,7 @@ class ProductRepository(Repository):
 
             cursor.execute(f"SELECT COUNT (*) FROM products;")
 
-            logger.info(f"Counting completed.")
+            logger.info("Counting completed.")
 
             return cursor.fetchall()[0][0]
 
@@ -162,6 +176,8 @@ def get_product_class(category: str) -> type(products.Product):
             product_class = products.Grain
         case _:
             product_class = products.Product
+            logger.warning("Invalid product category.",
+                           extra={"extra_parameters": {"category": category}})
 
     return product_class
 
@@ -171,24 +187,25 @@ def add_products(args: Namespace) -> None:
         products_list = file.read()
         products_list = json.loads(products_list)
 
+        repo = ProductRepository()
+
         for product_dict in products_list:
             product_class = get_product_class(product_dict["category"])
-            product = product_class(product_dict["name"],
-                                    product_dict.get("expiry_date")
-                                    or calculate_date(product_dict["freshness_in_days"]),
-                                    product_dict.get("quantity"))
+            if product_class != products.Product:
+                product = product_class(product_dict["name"],
+                                        product_dict.get("expiry_date")
+                                        or calculate_date(product_dict["freshness_in_days"]),
+                                        product_dict.get("quantity"))
 
-            repo = ProductRepository()
+                found_product = repo.get_by_name_and_date(product.name, product.expiry_date)
 
-            product_status = repo.search(conditions={"name": product.name, "expiry_date": product.expiry_date})
-
-            if product_status:
-                # product_status is one-element list of tuples
-                # the last element of the tuple is quantity value
-                product.quantity = product.quantity + product_status[0][-1]
-                repo.update(product)
-            else:
-                repo.add(product)
+                if found_product:
+                    # product_status is one-element list of tuples
+                    # the last element of the tuple is quantity value
+                    product.quantity = product.quantity + found_product[0][-1]
+                    repo.update(product)
+                else:
+                    repo.add(product)
 
 
 def remove_products(args: Namespace) -> None:
@@ -196,20 +213,20 @@ def remove_products(args: Namespace) -> None:
         products_list = file.read()
         products_list = json.loads(products_list)
 
+        repo = ProductRepository()
+
         for product_dict in products_list:
             product_id = product_dict["product_id"]
 
-            repo = ProductRepository()
-
             product = repo.get(product_id)
 
-            quantity_used = product_dict.get("quantity", product.DEFAULT_QUANTITY)
-
-            if product.quantity > quantity_used:
-                product.quantity = product.quantity - quantity_used
-                repo.update(product)
-            else:
-                repo.remove(product_id)
+            if product:
+                quantity_used = product_dict.get("quantity", product.DEFAULT_QUANTITY)
+                if product.quantity > quantity_used:
+                    product.quantity = product.quantity - quantity_used
+                    repo.update(product)
+                else:
+                    repo.remove(product_id)
 
 
 def display_products(args: Namespace) -> list[tuple]:
@@ -218,7 +235,6 @@ def display_products(args: Namespace) -> list[tuple]:
         request_body = json.loads(request_body)
 
         repo = ProductRepository()
-
         result = repo.search(conditions=request_body.get("conditions"), columns=request_body.get("columns", "*"),
                              limit=request_body.get("limit"))
